extract form steps into separate components
